+++ packages/webcomponents/src/components/payments-list/payments-list-core.tsx
@@ -1,20 +1,28 @@
 import { Component, h, Prop, State, Watch, Event, EventEmitter } from '@stencil/core';
-import { PagingInfo, Payment, pagingDefaults } from '../../api';
+import { PagingInfo, Payment, PaymentsParams, pagingDefaults } from '../../api';
 import { MapPaymentStatusToBadge, formatCurrency, formatDate, formatTime } from '../../utils/utils';
 import { ComponentError } from '../../api/ComponentError';
+import { tableExportedParts } from '../../ui-components/table/exported-parts';
 import { StyledHost, TableEmptyState, TableErrorState, TableLoadingState } from '../../ui-components';
+import { onFilterChange } from '../../ui-components/filters/utils';
 
 @Component({
-  tag: 'payments-list-core',
+  tag: 'payments-list-core'
 })
 export class PaymentsListCore {
-  @Prop() getPayments: Function;
-
   @State() payments: Payment[] = [];
   @State() loading: boolean = true;
   @State() errorMessage: string;
   @State() paging: PagingInfo = pagingDefaults;
-  @State() params: any;
+  @State() params: PaymentsParams = {};
+  
+  @Prop() getPayments: Function;
+
+  @Watch('params')
+  @Watch('getPayments')
+  updateOnPropChange() {
+    this.fetchData();
+  }
 
   @Event({
     eventName: 'payment-row-clicked',
@@ -22,12 +30,6 @@ export class PaymentsListCore {
   }) rowClicked: EventEmitter<Payment>;
 
   @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentError>;
-
-  @Watch('params')
-  @Watch('getPayments')
-  updateOnPropChange() {
-    this.fetchData();
-  }
 
   componentWillLoad() {
     if (this.getPayments) {
@@ -75,19 +77,28 @@ export class PaymentsListCore {
     this.rowClicked.emit(this.payments.find((payment) => payment.id === clickedPaymentID));
   };
 
+  setParamsOnChange = (name: string, value: string) => {
+    let newParams = { [name]: value };
+    this.params = onFilterChange(newParams, this.params);
+  }
+
+  clearParams = () => {
+    this.params = {};
+  }
+
   get entityId() {
     return this.payments.map((payment) => payment.id);
   }
 
   get columnData() {
     return [
-      ['Made On', 'The date and time each payment was made'],
+      ['Date', 'The date and time each payment was made'],
       ['Amount', 'The dollar amount of each payment'],
-      ['Description', 'The payment description, if you provided one'],
-      ['Cardholder', 'The name associated with the payment method'],
-      ['Payment Method', 'The brand and last 4 digits of the payment method'],
       ['Status', 'The current status of each payment'],
-      ['Payment ID', 'The unique identifier of each payment'],
+      ['Type', 'The type of each payment'],
+      ['Description', 'The payment description, if you provided one'],
+      ['Account Holder', 'The name associated with the payment method'],
+      ['Payment Method', 'The brand and last 4 digits of the payment method'],
     ];
   }
 
@@ -96,19 +107,19 @@ export class PaymentsListCore {
       {
         type: 'head',
         value: `
-          <div class='fw-bold'>${formatDate(payment.created_at)}</div>
-          <div class='fw-bold'>${formatTime(payment.created_at)}</div>
+          <div class="fw-bold">${formatDate(payment.created_at)}</div>
+          <div class="fw-bold">${formatTime(payment.created_at)}</div>
         `,
       },
-      formatCurrency(payment.amount),
-      payment.description,
-      payment.payment_method.payersName,
-      payment.payment_method.lastFourDigits,
+      formatCurrency(payment.amount, true, true),
       {
         type: 'inner',
         value: MapPaymentStatusToBadge(payment.status),
       },
-      payment.id,
+      payment.paymentType,
+      payment.description,
+      payment.payment_method.payersName,
+      payment.payment_method.lastFourDigits,
     ]);
   }
 
@@ -124,35 +135,20 @@ export class PaymentsListCore {
     return !this.showEmptyState && !this.showErrorState;
   }
 
-  handleDateChange = (name: string, value: string) => {
-    this.params = { ...this.params, [name]: value };
-  }
-
   render() {
     return (
-      <StyledHost exportparts="label">
-        <div class="row gy-3 mb-4">
-          <div class="col-2">
-            <form-control-date
-              name="created_after"
-              label="Start Date"
-              inputHandler={this.handleDateChange}
-            />
-          </div>
-          <div class="col-2">
-            <form-control-date
-              name="created_before"
-              label="End Date"
-              inputHandler={this.handleDateChange}
-            />
-          </div>
-        </div>
-        <div class="table-wrapper" part="table-wrapper">
-          <table class="table table-hover" part="table">
+      <StyledHost exportparts={tableExportedParts}>
+        <payments-list-filters 
+          params={this.params} 
+          setParamsOnChange={this.setParamsOnChange}
+          clearParams={this.clearParams}
+        />
+        <div class="table-wrapper">
+          <table class="table table-hover">
             <thead class="table-head sticky-top" part="table-head">
               <tr class="table-light text-nowrap" part="table-head-row">
                 {this.columnData?.map((column) => (
-                  <th part="table-head-cell" scope="col" title={Array.isArray(column) ? column[1] : ''}>
+                  <th part="table-head-cell" scope="col" title={Array.isArray(column) ? column[1] : ""}>
                     {!Array.isArray(column) ? column : column[0]}
                   </th>
                 ))}
@@ -177,7 +173,7 @@ export class PaymentsListCore {
                     data-test-id="table-row"
                     data-row-entity-id={this.entityId[index]}
                     onClick={this.rowClickHandler}
-                    part={`table-row ${index % 2 ? 'table-row-even' : 'table-row-odd'}`}
+                    part={`table-row ${index % 2 ? "table-row-even" : "table-row-odd"}`}
                   >
                     {data.map((dataEntry: any) => {
                       let nestedHtml = dataEntry?.type;